updated tests for cleanup

+++ sdk/test/IntegrationTests/Tests/SyncManager.cs
@@ -75,26 +75,31 @@ namespace AWSSDK_DotNet.IntegrationTests.Tests
                 FacebookUtilities.DeleteFacebookUser(facebookUser);
 #endif
             //drop all the tables from the db
-            using (SQLiteConnection connection = new SQLiteConnection(string.Format("Data Source={0};Version=3;", DB_FILE_NAME)))
+
+            var filePath = System.IO.Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "CognitoSync", DB_FILE_NAME);
+            if (File.Exists(filePath))
             {
-                connection.Open();
-
-                using (SQLiteCommand cmd = connection.CreateCommand())
+                using (SQLiteConnection connection = new SQLiteConnection(string.Format("Data Source={0};Version=3;", filePath)))
                 {
-                    cmd.CommandText = "DROP TABLE records";
-                    cmd.ExecuteNonQuery();
-                }
+                    connection.Open();
 
-                using (SQLiteCommand cmd = connection.CreateCommand())
-                {
-                    cmd.CommandText = "DROP TABLE datasets";
-                    cmd.ExecuteNonQuery();
-                }
+                    using (SQLiteCommand cmd = connection.CreateCommand())
+                    {
+                        cmd.CommandText = "DROP TABLE records";
+                        cmd.ExecuteNonQuery();
+                    }
 
-                using (SQLiteCommand cmd = connection.CreateCommand())
-                {
-                    cmd.CommandText = "DROP TABLE kvstore";
-                    cmd.ExecuteNonQuery();
+                    using (SQLiteCommand cmd = connection.CreateCommand())
+                    {
+                        cmd.CommandText = "DROP TABLE datasets";
+                        cmd.ExecuteNonQuery();
+                    }
+
+                    using (SQLiteCommand cmd = connection.CreateCommand())
+                    {
+                        cmd.CommandText = "DROP TABLE kvstore";
+                        cmd.ExecuteNonQuery();
+                    }
                 }
             }
         }
